icedtea6-native 1.7.5: fix build on old build hosts
Set ALT_FREETYPE_HEADERS_PATH so openjdk-ecj doesn't look in
/usr/include for freetype2 headers.

Build xawt sizer.32 on x86_64. A precompiled sizer.32 is included in
the icedtea tarball, but this could be an accident, and it doesn't
run on older build systems that can't handle DT_GNU_HASH executables;
this later causes the openjdk-6 build to fail for 32-bit targets.

Signed-off-by: Michael Smith <msmith@cbnco.com>
Cc: Henning Heinold <heinold@inf.fu-berlin.de>