opensp_1.5.bb: Fix compilation failure seen with -Os/gcc-4.5, Bugzilla/5494
This has been reported many times on mailing list and also in bugzilla/5494
The problem seems to be that gcc is not inlining the functions
marked inline in class definitions and emits calls to them but then
there is no definitions of these functions elsewhere. I dont know the
real cause yet could be a gcc problem.
Problem is when compiling

 armeb-oe-linux-uclibceabi-g++ -march=armv5te -mtune=xscale
-DHAVE_CONFIG_H -I. -I.. -I../include -I../generic
-isystem/scratch/oe/sysroots/armv5teb-oe-linux-uclibceabi/usr/include
-isystem/scratch/oe/sysroots/armv5teb-oe-linux-uclibceabi/usr/include
-fexpensive-optimizations -fomit-frame-pointer -frename-registers -Os
-pipe -fpermissive -fvisibility-inlines-hidden
-fvisibility-inlines-hidden -MT SgmlsEventHandler.o -MD -MP -MF
.deps/SgmlsEventHandler.Tpo -c -o SgmlsEventHandler.o
SgmlsEventHandler.cxx

Signed-off-by: Khem Raj <raj.khem@gmail.com>