udev: Remove duplicate udev firmware loading rule
A duplicate udev rule calling firmware.sh exists in
50-udev-default.rules and
udev.rules

This causes the kernel to sometimes get the wrong the firmware due to a
race between the two calls.

This is with udev124 from the stable2009 openembedded branch however a
quick
glance suggests that the same problem exists on the dev branch.

50-udev-default.rules comes from the upstream tarball
udev.rules is in the openembedded recipe.

This causes the infamous "Error number -110" while uploading firmware,
i.e:

[ 4171.357272] firmware: requesting zd1211/zd1211b_ub
[ 4171.375779] firmware: requesting zd1211/zd1211b_uphr
[ 4193.981056] firmware: requesting zd1211/zd1211b_ub
[ 4193.986433] firmware: requesting zd1211/zd1211b_uphr
[ 4195.033695] usb 8-3: USB control request for firmware upload failed.  Error number -110

* Fixes Bugzilla/5334

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Acked-by: Koen Kooi <koen@openembedded.org>